Remove required PHP versions
This is already handled by Magento repositories and when testing 2.2.6 and 2.3 Magento version seperately it requires that we add 7.2 when uprgrading to 2.3. Suggesting the removal of required PHP as we will need to request a new PHP version be added every few major Magento versions when Magento already handles this.